Added initializer lists for const class members

+++ Project1/Project1/src/Pointer.hpp
@@ -18,20 +18,18 @@ using namespace std;
 
 class Page{
 public:
-    Page(uint64_t address, unsigned int max, unsigned int size);
+    Page(uint64_t address, unsigned int max, size_t size);
     ~Page();
-    
     bool is_full();
     uint64_t get_next_address();
     void *get_reference_of(unsigned int offset);
-    
+
 private:
     void *memory;
     const size_t Node_size;
     const unsigned int max_count;
     unsigned int current_offset;
     const uint64_t root_address;
-    
 };
 
 
@@ -40,22 +38,17 @@ class MemoryAllocator{
 public:
     MemoryAllocator(int Node_size);
     ~MemoryAllocator();
-    
     uint64_t allocate();
     void* reference(uint64_t vortual_pointer);
-    
+
 private:
     const size_t Node_size;
     const unsigned int per_page_size;
     uint64_t next_page_address;
     vector<Page*>pages;
-    
+
     Page* addPage();
     unsigned int get_page_size(unsigned int Node_size);
 };
-
-
-
-
 
 #endif /* Pointer_hpp */
